refactor(backend): adopt src/sensor_api factory/blueprints/services layout
Restructure the backend to the CLAUDE.md section 1.3 target (#24):

- Move backend/app/ -> backend/src/sensor_api/ (src layout). Split the
  flat modules into a package: config.py, extensions.py (db, migrate),
  errors.py, sensors.py (domain), and blueprints/{sensors,health}/.
- Thin routes: sensors/ has routes (HTTP only), schemas (limit validation
  + serialization), services (query), models. The root / page folds into
  the health blueprint alongside /health and /ready.
- Factory does wiring only; three config classes unchanged.
- worker -> sensor_api/worker.py (python -m sensor_api.worker); run.py
  kept as a zero-config local entrypoint.
- Dockerfile: PYTHONPATH=/app/src, gunicorn sensor_api:create_app().
- compose: migrate runs flask --app sensor_api db upgrade; worker runs
  the module; worker healthcheck greps sensor_api.worker.
- pyproject.toml moves to backend/ (pythonpath=["src"]); CI backend job
  runs from backend/ (ruff check . / pytest).
- Tests import sensor_api.*; 26 pass. Relocated model is a no-op against
  the committed migration (verified: "No changes in schema detected").

Closes #24.
