+++ backend/src/sensor_api/blueprints/health/__init__.py
@@ -1,0 +1,3 @@
+from .routes import health
+
+__all__ = ["health"]
