+++ backend/src/sensor_api/blueprints/sensors/__init__.py
@@ -1,0 +1,4 @@
+from . import models  # noqa: F401  -- register model on metadata for Alembic
+from .routes import api
+
+__all__ = ["api"]
