+++ backend/src/sensor_api/blueprints/sensors/services.py
@@ -1,30 +1,10 @@
-from datetime import timezone
-
+# encoding: utf-8
 from sqlalchemy import select
 
-from .database import db
+from ...extensions import db
+from ...sensors import AnalogSensor, DiscreteSensor
 from .models import SensorData
-from .sensors import AnalogSensor, DiscreteSensor
-
-# Bounds for how many readings one query may return (inclusive). The API
-# boundary validates ?limit= against these; the service default is used when
-# no limit is supplied.
-DEFAULT_LIMIT = 100
-MIN_LIMIT = 1
-MAX_LIMIT = 100
-
-
-def _iso_utc(value):
-    """Serialize a datetime as an ISO-8601 string in UTC.
-
-    Stored timestamps are naive UTC (the DB ``current_timestamp`` default);
-    they are tagged as UTC so the wire format is unambiguous (``+00:00``).
-    """
-    if value is None:
-        return None
-    if value.tzinfo is None:
-        value = value.replace(tzinfo=timezone.utc)
-    return value.astimezone(timezone.utc).isoformat()
+from .schemas import DEFAULT_LIMIT
 
 
 class SensorService(object):
@@ -69,25 +49,13 @@ class SensorService(object):
 
     @staticmethod
     def fetch_data(limit=DEFAULT_LIMIT):
-        """Get the latest sensor readings, newest first.
+        """Return the latest sensor readings (ORM rows), newest first.
 
-        Returns a list of plain dicts with the timestamp serialized as an
-        ISO-8601 UTC string.
+        Serialization is the boundary's job (see schemas.serialize_reading).
         """
         statement = (
             select(SensorData)
             .order_by(SensorData.timestamp.desc())
             .limit(limit)
         )
-        rows = db.session.execute(statement).scalars().all()
-
-        return [
-            {
-                "id": row.id,
-                "timestamp": _iso_utc(row.timestamp),
-                "temperature": row.temperature,
-                "humidity": row.humidity,
-                "vibration": row.vibration,
-            }
-            for row in rows
-        ]
+        return db.session.execute(statement).scalars().all()
