+++ backend/src/sensor_api/errors.py
