--- pyproject.toml
@@ -1,10 +1,0 @@
-# Project-level config. Currently scoped to test discovery so the backend
-# test suite collects from a clean clone regardless of the working directory.
-#
-# The tests import `from backend.app import ...`, which requires the repo
-# root on sys.path. `pythonpath = ["."]` (relative to this file's directory,
-# the rootdir) makes `backend` importable without an editable install.
-
-[tool.pytest.ini_options]
-pythonpath = ["."]
-testpaths = ["backend/tests"]
